backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
dnspython==2.4.2
docopt==0.6.2
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.2
pipreqs==0.4.13
pymongo==4.6.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==68.2.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
wheel==0.41.3
yarg==0.1.9
